Improve IPA representation of "PuTTY", as suggested by John Lunney.  My rationale (as mailed to him):
I think you're right.  I got the pronunciation there from the second edition
of the OED and my Collins dictionary at home, both of which believe that
"pretty" is pronounced /'prItI/, but, at least to me, those two vowels are
different.  Both of them think that /i/ doesn't occur in English words, the
vowel in "beat" being /i:/.  The third edition of the OED, though, adds /i/
as an English vowel in its pronunciation guide, with "happy" as an example
of its use.  I'll update the FAQ following your suggestion.